+++ src/components/HandlerManager/HandlerManager.js
@@ -25,7 +25,7 @@ class HandlerManager extends React.Component {
   render() {
     const { stripes, data, props } = this.props;
     return (this.components.map(Component => (
-      <Component key={Component.name} stripes={stripes} data={data} {...props} />
+      <Component key={Component.name} stripes={stripes} actAs="handler" data={data} {...props} />
     )));
   }
 }